+++ frontend/src/help/Help.tsx
@@ -2,9 +2,8 @@ import { createSignal, JSX, lazy, Match, Show, Switch } from 'solid-js'
 import * as L from 'leaflet'
 import { bottomButtonsSize, keyWidth, soundingWidth, surfaceOverMap } from '../styles/Styles';
 import * as fakeData from './data';
-import { showDate } from '../data/ForecastMetadata';
+import { showDate, xcFlyingPotentialLayerName } from '../shared';
 import { type Domain } from '../State';
-import { xcFlyingPotentialLayer } from '../layers/ThQ';
 
 export const Help = (props: { domain: Domain }): JSX.Element => {
 
@@ -121,7 +120,7 @@ const MapHelp = (props: { domain: Domain }): JSX.Element => {
 
 const lazyMeteogram = (props: { domain: Domain }): JSX.Element =>
   lazy<() => JSX.Element>(() => {
-    return import(/* webpackPrefetch: true */ '../diagrams/Meteogram').then(module => {
+    return import('../diagrams/Meteogram').then(module => {
       const { key, view } = module.meteogram(fakeData.locationForecasts, props.domain.state);
       return { default: () => <>{ key }{ view }</> }
     })
@@ -144,7 +143,7 @@ const MeteogramHelp = (props: { domain: Domain }): JSX.Element => <>
     The <b>top row</b> (“XC?”) shows the estimated cross-country flying potential (between 0% and 100%).
     The higher the number, the higher the chances to fly long distances. It takes into account the
     boundary layer depth, the average thermal velocity, the wind speed, and the sunshine. Select the
-    layer “{ xcFlyingPotentialLayer.name }” in the map view to learn more about how it works.
+    layer “{ xcFlyingPotentialLayerName }” in the map view to learn more about how it works.
   </p>
   <p>
     The <b>second row</b> (“m/s”) shows the estimated average thermal velocity (in m/s) within the boundary layer.
@@ -199,7 +198,7 @@ const MeteogramHelp = (props: { domain: Domain }): JSX.Element => <>
 
 const lazySounding = (props: { domain: Domain }): JSX.Element =>
   lazy<() => JSX.Element>(() => {
-    return import(/* webpackPrefetch: true */ '../diagrams/Sounding').then(module => {
+    return import('../diagrams/Sounding').then(module => {
       const { key, view } = module.sounding(fakeData.detailedForecast, fakeData.groundLevel, true, props.domain.state);
       return { default: () => <>{ key }{ view }</> }
     })
